backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.41.0"}
fastapi = "^0.129.0"
sqlalchemy = "^2.0.46"
alembic = "^1.18.4"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
email-validator = "^2.3.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
argon2-cffi = "^25.1.0"
python-multipart = "^0.0.22"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
ruff = "^0.15.1"
black = "^26.1.0"
mypy = "^1.19.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
